requirements.txt
streamlit==1.22.0
# Upgraded TensorFlow and Keras for compatibility with keras.src
tensorflow-cpu==2.15.0
keras==3.3.3
protobuf==3.20.3
pandas==1.5.3
numpy==1.23.5
requests
Pillow
